samples-master/python/level1_single_api/9_amct/amct_pytorch/hif8_fp8_calibration/requirements.txt
torch==2.1.0+cpu
torch-npu==2.1.0.post3
torchvison==0.16.0+cpu
transformers==4.40.0
accelerate==0.30.1
datasets==2.19.1
sentencepiece==0.2.0
numpy==1.23.5
protobuf==3.13.0
Pillow==9.5.0